update kaa client for release 0.1.2

+++ package/utils/kaa_emu_application/src/src/kaa/kaa_platform_utils.h
@@ -1,17 +1,17 @@
-/*
- * Copyright 2014 CyberVision, Inc.
+/**
+ *  Copyright 2014-2016 CyberVision, Inc.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *       http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
  */
 
 
@@ -63,8 +63,8 @@ kaa_error_t kaa_platform_message_header_write(kaa_platform_message_writer_t* wri
                                             , uint16_t protocol_version);
 
 kaa_error_t kaa_platform_message_write_extension_header(kaa_platform_message_writer_t* writer
-                                                      , uint8_t extension_type
-                                                      , uint32_t options
+                                                      , uint16_t extension_type
+                                                      , uint16_t options
                                                       , uint32_t payload_size);
 
 
@@ -89,8 +89,8 @@ kaa_error_t kaa_platform_message_header_read(kaa_platform_message_reader_t* read
                                            , uint16_t *extension_count);
 
 kaa_error_t kaa_platform_message_read_extension_header(kaa_platform_message_reader_t *reader
-                                                     , uint8_t *extension_type
-                                                     , uint32_t *extension_options
+                                                     , uint16_t *extension_type
+                                                     , uint16_t *extension_options
                                                      , uint32_t *extension_payload_length);
 
 bool kaa_platform_message_is_buffer_large_enough(kaa_platform_message_reader_t *reader
